Move UserMatch component in a separate file
